Condense to one settings page and add navigation bar
