+++ react-app/src/pages/SettingsPage/SettingsPage.module.css
@@ -1,9 +1,20 @@
+.container {
+  height: 100vh;
+  width: 100%;
+}
+
+.rightContainer {
+  position: absolute;
+  width: 87%;
+  right: 0;
+}
+
 .inputs {
   flex-grow: 1;
   display: flex;
   flex-direction: column;
   box-sizing: border-box;
-  border: 1px solid black;
+  border: 1px solid #000000;
   border-radius: 20px;
   width: 60%;
   margin-left: 25%;
@@ -11,7 +22,7 @@
 }
 
 .settings {
-  margin-top: 150px;
+  margin-top: 100px;
 }
 
 .title {
@@ -24,22 +35,6 @@
   font-weight: 600;
   font-size: 25px;
   line-height: 38px;
-}
-
-.inputs {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-  border: 1px solid black;
-  border-radius: 20px;
-  width: 60%;
-  margin-left: 25%;
-  margin-right: 25%;
-}
-
-.settings {
-  margin-top: 100px;
 }
 
 .box {
@@ -101,32 +96,5 @@
 }
 
 .buttons {
-  padding-bottom: 30px;
-  padding-top: 30px;
-  padding-left: 5px;
-}
-
-.buttonLeft {
-  margin-top: 15px;
-  border-style: none;
-  background-color: var(--color-grey);
-  border-radius: 20px;
-  font-size: 1.4rem;
-  color: black;
-  width: 265px;
-  height: 61px;
-  margin-left: 25%;
-}
-
-.buttonRight {
-  margin-top: 15px;
-  border-style: none;
-  background-color: var(--color-grey);
-  border-radius: 20px;
-  font-size: 1.4rem;
-  color: black;
-  width: 265px;
-  height: 61px;
-  float: right;
-  margin-right: 15%;
+  display: flex;
 }
