--- react-app/src/pages/TeacherSettingsPage/TeacherSettingsPage.module.css
@@ -1,89 +1,0 @@
-.inputs {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-  border: 1px solid #000000;
-  border-radius: 20px;
-  width: 60%;
-  margin-left: 25%;
-  margin-right: 25%;
-}
-
-.settings {
-  margin-top: 100px;
-}
-
-.title {
-  text-align: center;
-  margin-left: 50%;
-  width: 105px;
-  height: 38px;
-
-  font-style: normal;
-  font-weight: 600;
-  font-size: 25px;
-  line-height: 38px;
-}
-
-.box {
-  border-bottom: 1px solid #000000;
-  padding-top: 30px;
-  padding-bottom: 25px;
-  padding-left: 5px;
-}
-
-.bottomBox {
-  padding-bottom: 30px;
-  padding-top: 30px;
-  padding-left: 5px;
-}
-
-.boxTitle {
-  font-style: normal;
-  font-weight: 400;
-  font-size: 1.5rem;
-  width: 200px;
-  padding-left: 20px;
-  display: inline-block;
-}
-
-.boxData {
-  float: center;
-  font-style: italic;
-  font-weight: 500;
-  font-size: 1.5rem;
-}
-
-.editKey {
-  background: none;
-  border: none;
-  outline: none;
-  float: right;
-  margin-right: 20px;
-  padding: 0px;
-  height: 32px;
-  color: #808080;
-  font-size: 2rem;
-  margin-right: 20px;
-  cursor: pointer;
-}
-
-.editButton {
-  background-color: var(--color-grey);
-  border-radius: 15px;
-  border-width: 0px;
-  width: 175px;
-  height: 40px;
-  outline: none;
-  float: right;
-  margin-right: 20px;
-  color: #000000;
-  font-weight: 400;
-  font-size: 0.9rem;
-  cursor: pointer;
-}
-
-.buttons {
-  display: flex;
-}
